requirements.txt
pandas
python-binance
requests
python-dotenv
influxdb-client==1.36.1
